Refactored getUserCreation & isContributionValid method to own files.
